Change params to ints

+++ src/parameters.h
@@ -8,8 +8,8 @@ struct Parameters {
 	double age_limit = 1;
 
 	int random_seed = 0;
-	uint8_t nodes = 1; // 0 for farkas, 1 for root node (> 0 for branching, support has been removed)
-	uint8_t num_threads = 1;
+	int nodes = 1; // 0 for farkas, 1 for root node (> 0 for branching, support has been removed)
+	int num_threads = 1;
 	bool show_output = true;
 	std::string column_retention = "low"; // Options: low, med, high
 };
